[refactor] Add more documentation, move authenticator with key type
