+++ crypto/authenticator.go
@@ -1,7 +1,6 @@
 package crypto
 
 import (
-	"crypto/ed25519"
 	"fmt"
 
 	"github.com/aptos-labs/aptos-go-sdk/bcs"
@@ -29,6 +28,15 @@ const (
 	AuthenticatorSingleSender AuthenticatorType = 4
 )
 
+// AuthenticatorImpl an implementation of an authenticator to provide generic verification across multiple types
+type AuthenticatorImpl interface {
+	bcs.Struct
+
+	// Verify Return true if this Authenticator approves
+	Verify(data []byte) bool
+}
+
+// Authenticator a generic authenticator type for a transaction
 type Authenticator struct {
 	Kind AuthenticatorType
 	Auth AuthenticatorImpl
@@ -38,6 +46,7 @@ func (ea *Authenticator) MarshalBCS(bcs *bcs.Serializer) {
 	bcs.Uleb128(uint32(ea.Kind))
 	ea.Auth.MarshalBCS(bcs)
 }
+
 func (ea *Authenticator) UnmarshalBCS(bcs *bcs.Deserializer) {
 	kindu := bcs.Uleb128()
 	if bcs.Error() != nil {
@@ -54,44 +63,9 @@ func (ea *Authenticator) UnmarshalBCS(bcs *bcs.Deserializer) {
 	}
 }
 
+// Verify verifies a message with the public key and signature
 func (ea *Authenticator) Verify(data []byte) bool {
 	return ea.Auth.Verify(data)
-}
-
-type AuthenticatorImpl interface {
-	bcs.Struct
-
-	// Verify Return true if this Authenticator approves
-	Verify(data []byte) bool
-}
-
-type Ed25519Authenticator struct {
-	PublicKey [ed25519.PublicKeySize]byte
-	Signature [ed25519.SignatureSize]byte
-}
-
-func (ea *Ed25519Authenticator) MarshalBCS(bcs *bcs.Serializer) {
-	bcs.WriteBytes(ea.PublicKey[:])
-	bcs.WriteBytes(ea.Signature[:])
-}
-func (ea *Ed25519Authenticator) UnmarshalBCS(bcs *bcs.Deserializer) {
-	kb := bcs.ReadBytes()
-	if len(kb) != ed25519.PublicKeySize {
-		bcs.SetError(fmt.Errorf("bad ed25519 public key, expected %d bytes but got %d", ed25519.PublicKeySize, len(kb)))
-		return
-	}
-	sb := bcs.ReadBytes()
-	if len(sb) != ed25519.SignatureSize {
-		bcs.SetError(fmt.Errorf("bad ed25519 signature, expected %d bytes but got %d", ed25519.SignatureSize, len(sb)))
-		return
-	}
-	copy(ea.PublicKey[:], kb)
-	copy(ea.Signature[:], sb)
-}
-
-// Verify Return true if the data was well signed
-func (ea *Ed25519Authenticator) Verify(data []byte) bool {
-	return ed25519.Verify(ed25519.PublicKey(ea.PublicKey[:]), data, ea.Signature[:])
 }
 
 // TODO: FeePayerAuthenticator, MultiAgentAuthenticator, MultiEd25519Authenticator, SingleSenderAuthenticator, SingleKeyAuthenticator
